Added player:remove_powerup() and player:give_powerup() + move player to entities_chars*
