+++ docs/generate.py
@@ -18,6 +18,7 @@ header_files = [
     "../src/game_api/entities_activefloors.hpp",
     "../src/game_api/entities_mounts.hpp",
     "../src/game_api/entities_monsters.hpp",
+    "../src/game_api/entities_chars.hpp",
     "../src/game_api/entities_items.hpp",
     "../src/game_api/entities_fx.hpp",
     "../src/game_api/entities_liquids.hpp",
@@ -37,6 +38,7 @@ api_files = [
     "../src/game_api/script/usertypes/save_context.cpp",
     "../src/game_api/script/usertypes/state_lua.cpp",
     "../src/game_api/script/usertypes/entity_lua.cpp",
+    "../src/game_api/script/usertypes/entities_chars_lua.cpp",
     "../src/game_api/script/usertypes/entities_floors_lua.cpp",
     "../src/game_api/script/usertypes/entities_activefloors_lua.cpp",
     "../src/game_api/script/usertypes/entities_mounts_lua.cpp",
