+++ src/game_api/entities_chars.hpp
@@ -1,0 +1,54 @@
+#pragma once
+
+#include "movable.hpp"
+#include <cstdint>
+
+class Player : public Movable
+{
+  public:
+    std::map<uint32_t, Entity*> powerups; // type id -> entity
+    Inventory* inventory_ptr;
+    Illumination* emitted_light;
+    int32_t linked_companion_child;  // entity uid
+    int32_t linked_companion_parent; // entity uid
+    size_t ai_func;
+    size_t input_ptr;
+    size_t p160;
+    int32_t i168;
+    int32_t i16c;
+    uint32_t jump_flags;
+    uint8_t some_timer;
+    uint8_t can_use;
+    uint8_t b176;
+    uint8_t b177;
+
+    void set_jetpack_fuel(uint8_t fuel);
+    uint8_t kapala_blood_amount();
+
+    /// Get the full name of the character, this will be the modded name not only the vanilla name.
+    std::u16string get_name();
+    /// Get the short name of the character, this will be the modded name not only the vanilla name.
+    std::u16string get_short_name();
+    /// Get the heart color of the character, this will be the modded heart color not only the vanilla heart color.
+    Color get_heart_color();
+    /// Check whether the character is female, will be `true` if the character was modded to be female as well.
+    bool is_female();
+
+    /// Set the heart color the character.
+    void set_heart_color(Color color);
+
+    /// Removes a currently applied powerup. Specify ENT_TYPE.ITEM_POWERUP_xxx, not ENT_TYPE.ITEM_PICKUP_xxx!
+    /// e.g. players[1]:remove_powerup(ENT_TYPE.POWERUP_CLIMBING_GLOVES)
+    /// Removing the Eggplant crown does not seem to undo the throwing of eggplants, the other powerups seem to work.
+    void remove_powerup(uint32_t powerup_type);
+
+    /// Gives the player the specified powerup. Specify ENT_TYPE.ITEM_POWERUP_xxx, not ENT_TYPE.ITEM_PICKUP_xxx!
+    void give_powerup(uint32_t powerup_type);
+};
+
+std::u16string get_character_name(int32_t type_id);
+std::u16string get_character_short_name(int32_t type_id);
+Color get_character_heart_color(int32_t type_id);
+bool is_character_female(int32_t type_id);
+
+void set_character_heart_color(int32_t type_id, Color color);
