+++ src/game_api/items.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "entities_chars.hpp"
 #include "entity.hpp"
 #include "state_structs.hpp"
