+++ src/game_api/movable.hpp
@@ -113,48 +113,6 @@ class Movable : public Entity
     std::uint32_t set_post_statemachine(std::function<void(Movable*)> post_state_machine);
 };
 
-class Player : public Movable
-{
-  public:
-    std::map<int64_t, int64_t> inside;
-    Inventory* inventory_ptr;
-    Illumination* emitted_light;
-    int32_t linked_companion_child;  // entity uid
-    int32_t linked_companion_parent; // entity uid
-    size_t ai_func;
-    size_t input_ptr;
-    size_t p160;
-    int32_t i168;
-    int32_t i16c;
-    uint32_t jump_flags;
-    uint8_t some_timer;
-    uint8_t can_use;
-    uint8_t b176;
-    uint8_t b177;
-
-    void set_jetpack_fuel(uint8_t fuel);
-    uint8_t kapala_blood_amount();
-
-    /// Get the full name of the character, this will be the modded name not only the vanilla name.
-    std::u16string get_name();
-    /// Get the short name of the character, this will be the modded name not only the vanilla name.
-    std::u16string get_short_name();
-    /// Get the heart color of the character, this will be the modded heart color not only the vanilla heart color.
-    Color get_heart_color();
-    /// Check whether the character is female, will be `true` if the character was modded to be female as well.
-    bool is_female();
-
-    /// Set the heart color the character.
-    void set_heart_color(Color color);
-};
-
-std::u16string get_character_name(int32_t type_id);
-std::u16string get_character_short_name(int32_t type_id);
-Color get_character_heart_color(int32_t type_id);
-bool is_character_female(int32_t type_id);
-
-void set_character_heart_color(int32_t type_id, Color color);
-
 class PlayerTracker
 {
   public:
