+++ src/game_api/script/usertypes/entities_chars_lua.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <sol/forward.hpp>
+
+class ScriptImpl;
+
+namespace NEntitiesChars
+{
+void register_usertypes(sol::state& lua, ScriptImpl* script);
+};
