+++ src/game_api/script/usertypes/entities_items_lua.cpp
@@ -99,6 +99,7 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
     lua["Entity"]["as_boomerang"] = &Entity::as<Boomerang>;
     lua["Entity"]["as_excalibur"] = &Entity::as<Excalibur>;
     lua["Entity"]["as_shield"] = &Entity::as<Shield>;
+    lua["Entity"]["as_powerup"] = &Entity::as<Powerup>;
 
     lua.new_usertype<Bomb>(
         "Bomb", "scale_hor", &Bomb::scale_hor, "scale_ver", &Bomb::scale_ver, "is_big_bomb", &Bomb::is_big_bomb, sol::base_classes, sol::bases<Entity, Movable>());
@@ -131,13 +132,6 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
     lua.new_usertype<Cape>("Cape", "floating_down", &VladsCape::floating_down, sol::base_classes, sol::bases<Entity, Movable>());
 
     lua.new_usertype<VladsCape>("VladsCape", "can_double_jump", &VladsCape::can_double_jump, sol::base_classes, sol::bases<Entity, Movable, Cape>());
-
-    lua.new_usertype<KapalaPowerup>(
-        "KapalaPowerup",
-        "amount_of_blood",
-        &KapalaPowerup::amount_of_blood,
-        sol::base_classes,
-        sol::bases<Entity, Movable>());
 
     lua.new_usertype<Mattock>(
         "Mattock",
@@ -805,6 +799,18 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
         sol::base_classes,
         sol::bases<Entity, Movable>());
 
+    lua.new_usertype<Powerup>(
+        "Powerup",
+        sol::base_classes,
+        sol::bases<Entity, Movable>());
+
+    lua.new_usertype<KapalaPowerup>(
+        "KapalaPowerup",
+        "amount_of_blood",
+        &KapalaPowerup::amount_of_blood,
+        sol::base_classes,
+        sol::bases<Entity, Movable, Powerup>());
+
     lua.new_usertype<ParachutePowerup>(
         "ParachutePowerup",
         "falltime_deploy",
@@ -814,14 +820,14 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
         "deploy",
         &ParachutePowerup::deploy,
         sol::base_classes,
-        sol::bases<Entity, Movable>());
+        sol::bases<Entity, Movable, Powerup>());
 
     lua.new_usertype<TrueCrownPowerup>(
         "TrueCrownPowerup",
         "timer",
         &TrueCrownPowerup::timer,
         sol::base_classes,
-        sol::bases<Entity, Movable>());
+        sol::bases<Entity, Movable, Powerup>());
 
     lua.new_usertype<AnkhPowerup>(
         "AnkhPowerup",
@@ -838,7 +844,7 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
         "music_on_off",
         &AnkhPowerup::music_on_off,
         sol::base_classes,
-        sol::bases<Entity, Movable>());
+        sol::bases<Entity, Movable, Powerup>());
 
     lua.new_usertype<YellowCape>(
         "YellowCape",
