+++ src/game_api/script/usertypes/entity_casting_lua.cpp
@@ -535,24 +535,24 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
     lua["TYPE_MAP"][538] = lua["Entity"]["as_rollingitem"];         // ITEM_PICKUP_TABLETOFDESTINY
     lua["TYPE_MAP"][539] = lua["Entity"]["as_rollingitem"];         // ITEM_PICKUP_SKELETON_KEY
     lua["TYPE_MAP"][541] = lua["Entity"]["as_playerbag"];           // ITEM_PICKUP_PLAYERBAG
-    lua["TYPE_MAP"][543] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_PASTE
-    lua["TYPE_MAP"][544] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_CLIMBING_GLOVES
-    lua["TYPE_MAP"][545] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_SPIKE_SHOES
-    lua["TYPE_MAP"][546] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_SPRING_SHOES
+    lua["TYPE_MAP"][543] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_PASTE
+    lua["TYPE_MAP"][544] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_CLIMBING_GLOVES
+    lua["TYPE_MAP"][545] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_SPIKE_SHOES
+    lua["TYPE_MAP"][546] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_SPRING_SHOES
     lua["TYPE_MAP"][547] = lua["Entity"]["as_kapalapowerup"];       // ITEM_POWERUP_KAPALA
-    lua["TYPE_MAP"][548] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_SPECTACLES
-    lua["TYPE_MAP"][549] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_PITCHERSMITT
-    lua["TYPE_MAP"][550] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_UDJATEYE
+    lua["TYPE_MAP"][548] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_SPECTACLES
+    lua["TYPE_MAP"][549] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_PITCHERSMITT
+    lua["TYPE_MAP"][550] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_UDJATEYE
     lua["TYPE_MAP"][551] = lua["Entity"]["as_parachutepowerup"];    // ITEM_POWERUP_PARACHUTE
-    lua["TYPE_MAP"][552] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_COMPASS
-    lua["TYPE_MAP"][553] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_SPECIALCOMPASS
-    lua["TYPE_MAP"][554] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_HEDJET
-    lua["TYPE_MAP"][555] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_CROWN
-    lua["TYPE_MAP"][556] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_EGGPLANTCROWN
+    lua["TYPE_MAP"][552] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_COMPASS
+    lua["TYPE_MAP"][553] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_SPECIALCOMPASS
+    lua["TYPE_MAP"][554] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_HEDJET
+    lua["TYPE_MAP"][555] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_CROWN
+    lua["TYPE_MAP"][556] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_EGGPLANTCROWN
     lua["TYPE_MAP"][557] = lua["Entity"]["as_truecrownpowerup"];    // ITEM_POWERUP_TRUECROWN
     lua["TYPE_MAP"][558] = lua["Entity"]["as_ankhpowerup"];         // ITEM_POWERUP_ANKH
-    lua["TYPE_MAP"][559] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_TABLETOFDESTINY
-    lua["TYPE_MAP"][560] = lua["Entity"]["as_movable"];             // ITEM_POWERUP_SKELETON_KEY
+    lua["TYPE_MAP"][559] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_TABLETOFDESTINY
+    lua["TYPE_MAP"][560] = lua["Entity"]["as_powerup"];             // ITEM_POWERUP_SKELETON_KEY
     lua["TYPE_MAP"][562] = lua["Entity"]["as_yellowcape"];          // ITEM_CAPE
     lua["TYPE_MAP"][563] = lua["Entity"]["as_vladscape"];           // ITEM_VLADS_CAPE
     lua["TYPE_MAP"][564] = lua["Entity"]["as_movable"];             // ITEM_PURCHASABLE_CAPE
@@ -614,24 +614,24 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
     lua["TYPE_MAP"][623] = lua["Entity"]["as_regenblock"];          // ACTIVEFLOOR_REGENERATINGBLOCK
     lua["TYPE_MAP"][626] = lua["Entity"]["as_movable"];             // FX_EGGSHIP_SHELL
     lua["TYPE_MAP"][627] = lua["Entity"]["as_movable"];             // FX_EGGSHIP_DOOR
-    lua["TYPE_MAP"][628] = lua["Entity"]["as_movable"];             // FX_EGGSHIP_CENTERJETFLAME
+    lua["TYPE_MAP"][628] = lua["Entity"]["as_movable"];             // FX_EGGSHIP_CENTERJETFLAME (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
     lua["TYPE_MAP"][629] = lua["Entity"]["as_movable"];             // FX_EGGSHIP_JETFLAME
     lua["TYPE_MAP"][630] = lua["Entity"]["as_movable"];             // FX_EGGSHIP_SHADOW
     lua["TYPE_MAP"][631] = lua["Entity"]["as_movable"];             // FX_SHADOW
-    lua["TYPE_MAP"][632] = lua["Entity"]["as_movable"];             // FX_TORNJOURNALPAGE
-    lua["TYPE_MAP"][633] = lua["Entity"]["as_movable"];             // FX_MAIN_EXIT_DOOR
-    lua["TYPE_MAP"][634] = lua["Entity"]["as_movable"];             // FX_BIRDIES
-    lua["TYPE_MAP"][635] = lua["Entity"]["as_movable"];             // FX_EXPLOSION
-    lua["TYPE_MAP"][636] = lua["Entity"]["as_movable"];             // FX_POWEREDEXPLOSION
-    lua["TYPE_MAP"][637] = lua["Entity"]["as_movable"];             // FX_MODERNEXPLOSION
-    lua["TYPE_MAP"][638] = lua["Entity"]["as_movable"];             // FX_OUROBORO_OCCLUDER
-    lua["TYPE_MAP"][639] = lua["Entity"]["as_movable"];             // FX_OUROBORO_HEAD
-    lua["TYPE_MAP"][640] = lua["Entity"]["as_movable"];             // FX_OUROBORO_TAIL
+    lua["TYPE_MAP"][632] = lua["Entity"]["as_movable"];             // FX_TORNJOURNALPAGE (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][633] = lua["Entity"]["as_movable"];             // FX_MAIN_EXIT_DOOR (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][634] = lua["Entity"]["as_movable"];             // FX_BIRDIES (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][635] = lua["Entity"]["as_movable"];             // FX_EXPLOSION (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][636] = lua["Entity"]["as_movable"];             // FX_POWEREDEXPLOSION (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][637] = lua["Entity"]["as_movable"];             // FX_MODERNEXPLOSION (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][638] = lua["Entity"]["as_movable"];             // FX_OUROBORO_OCCLUDER (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][639] = lua["Entity"]["as_movable"];             // FX_OUROBORO_HEAD (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][640] = lua["Entity"]["as_movable"];             // FX_OUROBORO_TAIL (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
     lua["TYPE_MAP"][641] = lua["Entity"]["as_movable"];             // FX_OUROBORO_TEXT
     lua["TYPE_MAP"][642] = lua["Entity"]["as_movable"];             // FX_OUROBORO_TRAIL
-    lua["TYPE_MAP"][643] = lua["Entity"]["as_movable"];             // ITEM_RUBBLE
-    lua["TYPE_MAP"][644] = lua["Entity"]["as_movable"];             // FX_COMPASS
-    lua["TYPE_MAP"][645] = lua["Entity"]["as_movable"];             // FX_SPECIALCOMPASS
+    lua["TYPE_MAP"][643] = lua["Entity"]["as_movable"];             // ITEM_RUBBLE (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][644] = lua["Entity"]["as_movable"];             // FX_COMPASS (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
+    lua["TYPE_MAP"][645] = lua["Entity"]["as_movable"];             // FX_SPECIALCOMPASS (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
     lua["TYPE_MAP"][646] = lua["Entity"]["as_movable"];             // FX_PLAYERINDICATOR
     lua["TYPE_MAP"][647] = lua["Entity"]["as_movable"];             // FX_PLAYERINDICATORPORTRAIT
     lua["TYPE_MAP"][648] = lua["Entity"]["as_movable"];             // FX_SALEICON
@@ -654,7 +654,7 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
     lua["TYPE_MAP"][665] = lua["Entity"]["as_movable"];             // FX_OLMECPART_SMALL
     lua["TYPE_MAP"][666] = lua["Entity"]["as_movable"];             // FX_OLMECPART_SMALLEST
     lua["TYPE_MAP"][667] = lua["Entity"]["as_movable"];             // FX_WEBBEDEFFECT
-    lua["TYPE_MAP"][668] = lua["Entity"]["as_movable"];             // FX_BUTTON
+    lua["TYPE_MAP"][668] = lua["Entity"]["as_movable"];             // FX_BUTTON (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
     lua["TYPE_MAP"][669] = lua["Entity"]["as_movable"];             // FX_BUTTON_DIALOG
     lua["TYPE_MAP"][670] = lua["Entity"]["as_movable"];             // FX_SLEEP_BUBBLE
     lua["TYPE_MAP"][671] = lua["Entity"]["as_movable"];             // FX_INK_BLINDNESS
@@ -731,7 +731,7 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
     lua["TYPE_MAP"][742] = lua["Entity"]["as_movable"];             // FX_EMPRESS
     lua["TYPE_MAP"][743] = lua["Entity"]["as_movable"];             // FX_EGGSHIP_HOOK_CHAIN
     lua["TYPE_MAP"][744] = lua["Entity"]["as_movable"];             // FX_MINIGAME_SHIP_DOOR
-    lua["TYPE_MAP"][745] = lua["Entity"]["as_movable"];             // FX_MINIGAME_SHIP_CENTERJETFLAME
+    lua["TYPE_MAP"][745] = lua["Entity"]["as_movable"];             // FX_MINIGAME_SHIP_CENTERJETFLAME (NOT IMPLEMENTED YET, FORCED TO MOVABLE)
     lua["TYPE_MAP"][746] = lua["Entity"]["as_movable"];             // FX_MINIGAME_SHIP_JETFLAME
     lua["TYPE_MAP"][748] = lua["Entity"]["as_movable"];             // FX_ANKH_ROTATINGSPARK
     lua["TYPE_MAP"][749] = lua["Entity"]["as_movable"];             // FX_ANKH_FALLINGSPARK
@@ -846,7 +846,7 @@ void register_usertypes(sol::state& lua, ScriptImpl* script)
     // lua["TYPE_MAP"][863] = lua["Entity"]["?"]; // LOGICAL_CAMERA_FLASH (UNKNOWN IN PLUGIN)
     // lua["TYPE_MAP"][864] = lua["Entity"]["?"]; // LOGICAL_ROOM_LIGHT (UNKNOWN IN PLUGIN)
     // lua["TYPE_MAP"][865] = lua["Entity"]["?"]; // LOGICAL_LIMB_ANCHOR (UNKNOWN IN PLUGIN)
-    // lua["TYPE_MAP"][866] = lua["Entity"]["?"]; // LOGICAL_PORTAL (UNKNOWN IN PLUGIN)
+    // lua["TYPE_MAP"][866] = lua["Entity"]["as_portal"];  // LOGICAL_PORTAL (NOT IMPLEMENTED YET)
     // lua["TYPE_MAP"][867] = lua["Entity"]["?"]; // LOGICAL_STATICLAVA_SOUND_SOURCE (UNKNOWN IN PLUGIN)
     // lua["TYPE_MAP"][868] = lua["Entity"]["?"]; // LOGICAL_STREAMLAVA_SOUND_SOURCE (UNKNOWN IN PLUGIN)
     // lua["TYPE_MAP"][869] = lua["Entity"]["?"]; // LOGICAL_STREAMWATER_SOUND_SOURCE (UNKNOWN IN PLUGIN)
